* commit '3d51a25ffb6bf07876f6c6736fe912f973d3ac9e':
  fix the bad merge
  Update utils/mem_mgr.cc
  Update utils/file_util.cc
  resolving merge issues
  fsync the destination directory after file copy
  newt change to build the package
  changes made while reading the code to identify the rss mem usage
  changes made while reading the code to identify the rss mem usage

+++ utils/socket_pool.h
@@ -23,7 +23,7 @@ class SocketPool {
  private:
   std::string hostname_;
   int port_;
-  int num_sockets_;
+  size_t num_sockets_;
 
   Sockaddr sockaddr_;
   std::vector<Socket*> sockets_;
